update-> restructured file directories for clear separation between admin and user
