+++ app/(admin)/admin/components/create-equipment.tsx
