+++ app/(admin)/admin/components/update-equipment.tsx
