+++ app/(admin)/admin_incident/page.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { supabase } from '@/config/supabaseClient'
 import { revalidatePath } from "next/cache";
-import { IncidentDataTable } from '@/app/admin_incident/components/incident-data-table'
-import { Incident, incident_columns } from "@/app/admin_incident/components/incident-columns"
+import { IncidentDataTable } from '@/app/(admin)/admin_incident/components/incident-data-table'
+import { Incident, incident_columns } from "@/app/(admin)/admin_incident/components/incident-columns"
 
 const Incidentreport = async () => {
   revalidatePath("/admin_incident")
