+++ app/(user)/equipmentpage/cart.tsx
