+++ app/(user)/equipmentpage/page.tsx
